+++ bot/events/messageCreate.js
@@ -1,7 +1,9 @@
-const config = require('../../utils/assets/json/_config/config.json');
 const { errorhandler } = require('../../utils/functions/errorhandler/errorhandler');
 const { getLang } = require('../../utils/functions/getData/getLang');
-const database = require('../db/db');
+const { getIgnoreMode } = require('../../utils/functions/getData/getIgnoreMode');
+const { isUserPremium } = require('../../utils/functions/premium/premium');
+const { delay } = require('../../utils/functions/delay/delay');
+const { getPrefix } = require('../../utils/functions/getData/getPrefix');
 
 async function messageCreate(message, bot) {
     if (message.author.bot) return;
@@ -13,27 +15,48 @@ async function messageCreate(message, bot) {
     var cmd = messageArray[0];
     var args = messageArray.slice(1);
 
-    const prefix = await database.query(`SELECT prefix FROM ${config.tables.mido_config} WHERE guild_id = ?`, message.guild.id)
-    .then(res => {
-        if(res === 0) {
-            return config.defaultprefix;
-        }else {
-            return res[0].prefix;
-        }
-    }).catch(async err => {
-        const lang = require(`../../../utils/assets/json/language/${await getLang(message.guild.id)}.json`)
-
-        errorhandler(err, lang.errors.general, message.channel);
-        return false;
-    });
-
-    if(!prefix) return;
+    const prefix = await getPrefix({
+        guild_id: message.guild.id
+    })
 
     if (cmd.startsWith(prefix)) {
+
+        const isIgnoreMode = await getIgnoreMode();
+
+        if(isIgnoreMode.error) {
+            errorhandler(isIgnoreMode.message);
+            const lang = require(`../../utils/assets/json/language/${await getLang(message.guild.id)}.json`)
+            message.reply({
+                content: lang.errors.general
+            }).then(async msg => {
+                await delay(5000);
+                msg.delete().catch(err => {}).catch(err => {})
+            }).catch(err => {})
+        }
+    
+        if(isIgnoreMode.ignoreMode) {
+            const lang = require(`../../utils/assets/json/language/${await getLang(message.guild.id)}.json`)
+            return message.reply({
+                content: lang.errors.ignoreModeOn
+            }).then(async msg => {
+                await delay(5000);
+                msg.delete().catch(err => {})
+            }).catch(err => {})
+        }
+
         let commandfile = bot.commands.get(cmd.slice(prefix.length));
 
         if (commandfile) { //&& blacklist(0, message)
-            return commandfile.run(bot, message, args);
+            
+            const isPremium = isUserPremium({
+                user_id: message.author.id
+            });
+
+            if(isPremium.error) return message.reply({
+                content: isPremium.message
+            });
+
+            return commandfile.run(bot, message, args, isPremium.premium, isPremium.platin);
         } else return;
 
     }